+++ packages/lit-analyzer/src/analyze/parse/parse-html-data/html-tag.ts
@@ -132,7 +132,7 @@ export function documentationForHtmlTag(htmlTag: HtmlTag, options: DescriptionOp
 		desc += `\n\n${descriptionList(
 			"Events",
 			items,
-			event => `${descriptionHeader(`@event ${event.name}`, 0, options)}${event.description ? ` - ${event.description}` : ""}`,
+			event => `${descriptionHeader(`@fires ${event.name}`, 0, options)}${event.description ? ` - ${event.description}` : ""}`,
 			options
 		)}`;
 	}